+ GameObject  - 달리기 스킬 추가  - 순찰 중에 공격에 돌입하고 제자리로 돌아갈 떄 생기던 순간이동 버그 수정  - 적군에게 잠재우기 스킬 추가    (잠재워질 경우는 스킬이 아니고 상태이기 때문에 SkillController가 아닌     GameObject 클래스에 넣게 되었음.)  - hide_update() 제거 및 SkillController로 이전 + HUDCollection  - 각 상태들 추가에 따른 Head Up Display 추가  - 달리기 상태, 은신 상태, 수면 상태 UI 추가 + ObjectController  - 스테이지에 생성된 객체의 스텟을 임의 조정 가능하도록 루아 함수에    연결시킴 + SkillController  - 달리기, 은신 상태 업데이트 루틴 추가 + StageManager.cpp  - 플레이어 객체가 죽고 패배하여 메인화면으로 돌아간 뒤, 다시 시작하면    스테이지 초기 메시지가 띄워질 때 동작 중지가 되지 않던 문제 수정 + BD_Button  - setButtonImage() 함수 추가. UIComponent 추상 객체에서 이미지를    게임 오브젝트의 스킬 이미지로 바꾸기가 기존 구조 상으로는 힘든 이유로    dynamic_cast를 통하여 위 함수 접근 가능하게끔 함. + BD_PopupMessage  - CCLabelTTF 클래스로부터 성능 향상을 위해 CCLabelBMFont로 변경
=> 그 외에 루아 스크립트 파일 일부 수정(시험용 스테이지 구성중.)

+++ Classes/Objects/CurveThrowObject.cpp
@@ -109,7 +109,7 @@ void CurveThrowObject::Destroy()
 	vector_clear(throwing_line);
 	vector_clear(effect_frame_name_list);
 }
-void CurveThrowObject::change_PosList_OnScrolling(float delta_X)
+void CurveThrowObject::change_PosList_OnScrolling(int delta_X)
 {
 	unsigned int throw_pos_size = throwing_line.size();
 	for(unsigned int i = 0; i < throw_pos_size; ++i)
